requirements.txt
# ApnaPG Backend Dependencies
# =============================

# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Database
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10

# Authentication (Clerk JWT)
pyjwt[crypto]==2.10.1

# Media Storage
cloudinary==1.42.1

# PDF Generation
reportlab==4.3.1

# Configuration
python-dotenv==1.1.0

# CORS & Utilities
python-multipart==0.0.20
httpx==0.28.1
